Add files via upload
The two main papers that the vision pipeline is based on. The Fast Feature Pyramids for Object detection describes a method to efficiently compute features at a large number of different scales. 

Filtered Channel Features for Pedestrian Detection is a recent paper that set a benchmark on real-time pedestrian detection, based partial on methods from the feature pyramid paper.